Parse all plain timestamps in a heading
Previously, plain timestamps (active or inactive) in a note content were
ignored.  The only timestamps we parsed were the scheduled, deadline, or closed
ones in the "header" of a heading.

This commit adds `timestamps`, a list of OrgRange in OrgHead, and modifies
`OrgSaxyParser` to collect plain timestamps.

This commit also adds tests to make sure plain timestamps are collected
correctly.

We assume that we do not care about plain timestamps occurring in a heading's
"header".